@@ -5,11 +5,11 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 27 // updated from 26 - AG
+    compileSdkVersion 29 // updated from 26 - AG
     defaultConfig {
         applicationId "com.example.grycz.imageprocessor"
         minSdkVersion 21
-        targetSdkVersion 27 // updated from 26 - AG
+        targetSdkVersion 29 // updated from 26 - AG
         multiDexEnabled true
         // added by AG 15.03.2018 in order to fix "error converting bytecode to dex multiple dex files define" error
         versionCode 1
@@ -32,19 +32,19 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
-    // implementation 'com.android.support:appcompat-v7:26.1.0' // old
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-//     implementation 'com.android.support:design:26.1.0' // old
-    implementation 'com.android.support:design:27.1.1'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.71"
+    implementation "androidx.appcompat:appcompat:1.1.0"
+    // For loading and tinting drawables on older versions of the platforml
+    implementation "androidx.appcompat:appcompat-resources:1.1.0"
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    implementation 'org.apache.httpcomponents:httpcore:4.4.1'
+    androidTestImplementation 'com.android.support.test:runner:1.0.2'
+    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'org.apache.httpcomponents:httpcore:4.4.10'
 //    compile 'com.example.grycz.imagecropper.Android-Image-Cropper.cropper.src.main.java.com.theartofdev.edmodo:android-image-cropper:2.6.+'
-    implementation 'com.theartofdev.edmodo:android-image-cropper:2.6.0'
-    implementation 'com.android.support:support-v4:27.1.1'
+    api 'com.theartofdev.edmodo:android-image-cropper:2.6.0'
+    implementation 'androidx.localbroadcastmanager:localbroadcastmanager:1.0.0'
+    implementation 'com.google.android.material:material:1.1.0'
 
 }
